+++ src/basic/components/common/index.ts
@@ -1,2 +1,1 @@
-export * from './SearchBar';
 export * from './EmptyState';
