+++ src/basic/components/product/index.ts
@@ -3,3 +3,4 @@ export * from './ProductPrice';
 export * from './ProductStockStatus';
 export * from './ProductDiscountBadge';
 export * from './ProductDiscountInfo';
+export * from './ProductGrid';
